Finished CP1

See merge request smigapav/era_project!5

+++ src/main/kotlin/cz/cvut/fel/ear/ear_project/service/UserService.kt
@@ -1,11 +1,7 @@
 package cz.cvut.fel.ear.ear_project.service
 
-import cz.cvut.fel.ear.ear_project.dao.PermissionsRepository
-import cz.cvut.fel.ear.ear_project.dao.ProjectRepository
 import cz.cvut.fel.ear.ear_project.dao.TaskRepository
 import cz.cvut.fel.ear.ear_project.dao.UserRepository
-import cz.cvut.fel.ear.ear_project.exceptions.ItemAlreadyPresentException
-import cz.cvut.fel.ear.ear_project.exceptions.ItemNotFound
 import cz.cvut.fel.ear.ear_project.model.Task
 import cz.cvut.fel.ear.ear_project.model.User
 import org.springframework.beans.factory.annotation.Autowired
@@ -18,41 +14,60 @@ class UserService(
     private val userRepository: UserRepository,
     @Autowired
     private val taskRepository: TaskRepository,
-    @Autowired
-    private val projectRepository: ProjectRepository,
-    @Autowired
-    private val permissionRepository: PermissionsRepository
 ) {
-    @Transactional
-    fun insertUser(user: User) {
+    fun createUser(name: String): User {
+        val user = User()
+        user.username = name
         userRepository.save(user)
+        return user
     }
 
-    @Transactional
     fun removeUser(user: User) {
+        if (!userExists(user)) {
+            throw IllegalArgumentException("User does not exist")
+        }
         userRepository.delete(user)
     }
 
     @Transactional
-    fun addTask(user: User, task: Task) {
+    fun addTask(
+        user: User,
+        task: Task,
+    ): User {
+        if (!userExists(user) || !taskExists(task)) {
+            throw IllegalArgumentException("User or task does not exist")
+        }
         user.addTask(task)
+        task.user = user
         userRepository.save(user)
-        task.addUser(user)
         taskRepository.save(task)
+        return user
     }
 
     @Transactional
-    fun removeTask(user: User, task: Task) {
-        user.tasks.remove(task)
+    fun removeTask(
+        user: User,
+        task: Task,
+    ) {
+        if (!userExists(user) || !taskExists(task)) {
+            throw IllegalArgumentException("User or task does not exist")
+        }
+        user.removeTask(task)
         userRepository.save(user)
         task.user = null
         taskRepository.save(task)
     }
 
-    @Transactional
-    fun changeUsername(user: User, username: String) {
+    fun changeUsername(
+        user: User,
+        username: String,
+    ): User {
+        if (!userExists(user)) {
+            throw IllegalArgumentException("User does not exist")
+        }
         user.username = username
         userRepository.save(user)
+        return user
     }
 
     fun findAllUsers(): List<User> {
@@ -61,5 +76,13 @@ class UserService(
 
     fun findUserById(id: Long): User? {
         return userRepository.findById(id).orElse(null)
+    }
+
+    fun userExists(user: User): Boolean {
+        return user.id != null && !userRepository.findById(user.id!!).isEmpty
+    }
+
+    fun taskExists(task: Task): Boolean {
+        return task.id != null && !taskRepository.findById(task.id!!).isEmpty
     }
 }
